+++ tests/integration/test_dataset_update.py
@@ -5,7 +5,7 @@ import pytest
 
 from bulk_data_service.checker import checker_run
 from helpers.helpers import get_and_clear_up_context  # noqa: F401
-from helpers.helpers import (
+from helpers.data_helpers import (
     check_last_known_good_dataset_values_are_set,
     check_last_known_good_dataset_values_are_unset,
     check_most_recent_get_attempt_downloaded_but_non_iati,
